requirements.txt
loguru
pydash
xxhash
pyyaml
pybind11
pydantic
setuptools
flatbuffers
alive-progress